Code base refactoring
